libhwc2.1: Ignore transparent layer
By the design of SurfaceFlinger cache, the layer with full alpha value
can be passed to HWC with the meaning of a cached layer. The patch in
HWC puts those layers into ignore layer stacks correspondingly.

[borbelyvince]:
Modify setLayerPlaneAlpha() so that geometry updates are registered.

Since we have ExynosDevice::validateDisplay instead of
ExynosDisplay::validateDisplay on pixel, call checkIgnoreLayers()
there.

Bug: 176560197
Test: Check ignore layers can work with layer alpha
Signed-off-by: HyunKyung Kim <hk310.kim@samsung.com>
Change-Id: I39bebed25c5623e2bbf83ff5c0b3c942302664c4